[FIX] hr_payroll_timesheet: Fixes for a database that has gone through multiple Odoo Upgrades
Init will run on install, even if the module existed in the prior version.  Because these init scripts are intended to clear (and maybe even re-name back on actual init of data), records we should be more generic with the intended name.

Additionally, if your database already had `hr_payroll_timesheet` based 'TS' work entry codes, then we cannot delete it if it is used on records like payslips (as a work type).  Delete the link to simply stop using the old record, but leave it behind to keep it working.