[FIX] connector_opencart: find orders by modified date instead of ID
Opencart will not always generate a new order id, and may update the date_added but not actually be ready for import.  To mitigate, we're now looking for orders modified after a set datetime.

Additionally, offer mechanisms to conver datetime to/from Opencart's timezone. This is used during the above import behavior, as well as the `date_order` field in Odoo.