[IMP] delivery_hibou: add get_to_ship_picking_packages utility
I find that this must be in the pre-amble of iteration on pickings to ensure they have the correct output when package carriers are in play.

Specifically, we need to be able to 're-enter' the send to shipper method with a new package for a different carrier and not even try to ship if all packages have tracking numbers.